The file Texml/handler.py is now namespace aware.
The class glue_handler can either be used as namespace aware or
not namespace aware. Whether it is namespace aware is handled
outside of this class.

If glue_handler is not namespace aware, it acts just as id did
before.

If glue_handler in in namespace aware mode, it will process elements
(including characters) in the namespace
http://getfo.sourceforge.net/texml/ns1 and ignore any elements.

The file scripts/texml.py now works with the option --ns. If this
options is chosen, the script becomes namespace aware and the
Texml/handler.py is invoked in namespace aware mode.


git-svn-id: svn://svn.code.sf.net/p/getfo/code/trunk@432 c9bd6c6d-8c10-4832-ae80-b1add196a410