Texml/handler.py ----------------
glue_handler no processes all elements with awareness of namspaces.
If the element in fact has no nampace (None), it is processed.
If the element has the wrong namespace, the invalid_xml method
from the handler class is invoked, and the script exits.

The method invalid_xml was added to the handler class. The script now
quits with invalid XML with a more informative message rather
than just raising a generic python error

scripts/texml.py
----------------

The option --ns is no longer valid, since namespace is used
automatically.

If the XML document is invalid, the script quits with an
informative message rather than throwing a generic python exception.

tests/Makefile
--------------

the script in scripts/texml.py is now used to run tests.


git-svn-id: svn://svn.code.sf.net/p/getfo/code/trunk@436 c9bd6c6d-8c10-4832-ae80-b1add196a410